Update RealSense setup logic to validate device is accessible
