+++ src/robotics_utils/visualization/__init__.py
@@ -1,0 +1,4 @@
+"""Import classes and functions for visualization."""
+
+from .display_images import Displayable as Displayable
+from .display_images import display_image as display_image
